fix: provision artist analytics page
